feat: delete lobby
