+++ client/package.json
@@ -27,7 +27,9 @@
     "ramda": "^0.28.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "sass": "^1.57.1",
     "socket.io-client": "^4.5.4",
+    "three-dots": "^0.3.2",
     "typescript": "4.9.4",
     "zustand": "^4.2.0"
   },
