+++ client/pages/_app.tsx
@@ -6,6 +6,7 @@ import CssBaseline from '@mui/material/CssBaseline';
 import { CacheProvider, EmotionCache } from '@emotion/react';
 import { dotGothicFont, theme } from '../src/theme';
 import createEmotionCache from '../src/createEmotionCache';
+import './style.scss';
 
 // Client-side cache, shared for the whole session of the user in the browser.
 const clientSideEmotionCache = createEmotionCache();
