+++ client/pages/style.scss
@@ -1,0 +1,1 @@
+@use 'three-dots';
